nk3: Move test into trussed module
