+++ pynitrokey/cli/trussed/tests.py
@@ -9,127 +9,26 @@
 # copied, modified, or distributed except according to those terms.
 
 import logging
-import platform
-import sys
-from dataclasses import dataclass
-from enum import Enum, auto, unique
 from hashlib import sha256
 from struct import unpack
 from threading import Thread
-from types import TracebackType
-from typing import Any, Callable, Iterable, Optional, Tuple, Type, Union
+from typing import Any, Optional
 
 from tqdm import tqdm
 
-from pynitrokey.cli.exceptions import CliException
-from pynitrokey.fido2 import device_path_to_str
+from pynitrokey.cli.trussed.test import TestContext, TestResult, TestStatus, test_case
 from pynitrokey.fido2.client import NKFido2Client
 from pynitrokey.helpers import local_print
-from pynitrokey.nk3.device import Nitrokey3Device
 from pynitrokey.nk3.utils import Fido2Certs
 from pynitrokey.trussed.base import NitrokeyTrussedBase
+from pynitrokey.trussed.device import NitrokeyTrussedDevice
 from pynitrokey.trussed.utils import Uuid, Version
 
 logger = logging.getLogger(__name__)
 
 
-TEST_CASES = []
-
 AID_ADMIN = [0xA0, 0x00, 0x00, 0x08, 0x47, 0x00, 0x00, 0x00, 0x01]
 AID_PROVISIONER = [0xA0, 0x00, 0x00, 0x08, 0x47, 0x01, 0x00, 0x00, 0x01]
-
-DEFAULT_EXCLUDES = ["bootloader", "provisioner"]
-
-
-ExcInfo = Tuple[Type[BaseException], BaseException, TracebackType]
-
-
-class TestContext:
-    def __init__(self, pin: Optional[str]) -> None:
-        self.pin = pin
-        self.firmware_version: Optional[Version] = None
-
-
-@unique
-class TestStatus(Enum):
-    SKIPPED = auto()
-    SUCCESS = auto()
-    FAILURE = auto()
-
-
-class TestResult:
-    def __init__(
-        self,
-        status: TestStatus,
-        data: Optional[str] = None,
-        exc_info: Union[ExcInfo, Tuple[None, None, None]] = (None, None, None),
-    ) -> None:
-        self.status = status
-        self.data = data
-        self.exc_info = exc_info
-
-
-TestCaseFn = Callable[[TestContext, NitrokeyTrussedBase], TestResult]
-
-
-class TestCase:
-    def __init__(self, name: str, description: str, fn: TestCaseFn) -> None:
-        self.name = name
-        self.description = description
-        self.fn = fn
-
-
-def test_case(name: str, description: str) -> Callable[[TestCaseFn], TestCaseFn]:
-    def decorator(func: TestCaseFn) -> TestCaseFn:
-        TEST_CASES.append(TestCase(name, description, func))
-        return func
-
-    return decorator
-
-
-def filter_test_cases(
-    test_cases: list[TestCase], names: Iterable[str]
-) -> Iterable[TestCase]:
-    for test_case in test_cases:
-        if test_case.name in names:
-            yield test_case
-
-
-@dataclass
-class TestSelector:
-    only: Iterable[str] = ()
-    all: bool = False
-    include: Iterable[str] = ()
-    exclude: Iterable[str] = ()
-
-    def select(self) -> list[TestCase]:
-        if self.only:
-            return list(filter_test_cases(TEST_CASES, self.only))
-
-        selected = []
-        for test_case in TEST_CASES:
-            if test_case.name in self.include:
-                selected.append(test_case)
-            elif test_case.name not in self.exclude:
-                if self.all or test_case.name not in DEFAULT_EXCLUDES:
-                    selected.append(test_case)
-        return selected
-
-
-def log_devices() -> None:
-    from fido2.hid import CtapHidDevice
-
-    ctap_devices = [device for device in CtapHidDevice.list_devices()]
-    logger.info(f"Found {len(ctap_devices)} CTAPHID devices:")
-    for device in ctap_devices:
-        descriptor = device.descriptor
-        path = device_path_to_str(descriptor.path)
-        logger.info(f"- {path} ({descriptor.vid:x}:{descriptor.pid:x})")
-
-
-def log_system() -> None:
-    logger.info(f"platform: {platform.platform()}")
-    logger.info(f"uname: {platform.uname()}")
 
 
 @test_case("uuid", "UUID query")
@@ -143,7 +42,7 @@ def test_uuid_query(ctx: TestContext, device: NitrokeyTrussedBase) -> TestResult
 def test_firmware_version_query(
     ctx: TestContext, device: NitrokeyTrussedBase
 ) -> TestResult:
-    if not isinstance(device, Nitrokey3Device):
+    if not isinstance(device, NitrokeyTrussedDevice):
         return TestResult(TestStatus.SKIPPED)
     version = device.admin.version()
     ctx.firmware_version = version
@@ -152,7 +51,7 @@ def test_firmware_version_query(
 
 @test_case("status", "Device status")
 def test_device_status(ctx: TestContext, device: NitrokeyTrussedBase) -> TestResult:
-    if not isinstance(device, Nitrokey3Device):
+    if not isinstance(device, NitrokeyTrussedDevice):
         return TestResult(TestStatus.SKIPPED)
     firmware_version = ctx.firmware_version or device.admin.version()
     if firmware_version.core() < Version(1, 3, 0):
@@ -186,7 +85,7 @@ def test_device_status(ctx: TestContext, device: NitrokeyTrussedBase) -> TestRes
 def test_bootloader_configuration(
     ctx: TestContext, device: NitrokeyTrussedBase
 ) -> TestResult:
-    if not isinstance(device, Nitrokey3Device):
+    if not isinstance(device, NitrokeyTrussedDevice):
         return TestResult(TestStatus.SKIPPED)
     if device.admin.is_locked():
         return TestResult(TestStatus.SUCCESS)
@@ -360,7 +259,7 @@ SE050_STEPS = [
 def test_se050(ctx: TestContext, device: NitrokeyTrussedBase) -> TestResult:
     from queue import Queue
 
-    if not isinstance(device, Nitrokey3Device):
+    if not isinstance(device, NitrokeyTrussedDevice):
         return TestResult(TestStatus.SKIPPED)
 
     que: Queue[Optional[bytes]] = Queue()
@@ -431,7 +330,7 @@ def test_se050(ctx: TestContext, device: NitrokeyTrussedBase) -> TestResult:
 
 @test_case("fido2", "FIDO2")
 def test_fido2(ctx: TestContext, device: NitrokeyTrussedBase) -> TestResult:
-    if not isinstance(device, Nitrokey3Device):
+    if not isinstance(device, NitrokeyTrussedDevice):
         return TestResult(TestStatus.SKIPPED)
 
     # drop out early, if pin is needed, but not provided
@@ -539,61 +438,3 @@ def test_fido2(ctx: TestContext, device: NitrokeyTrussedBase) -> TestResult:
     )
 
     return TestResult(TestStatus.SUCCESS)
-
-
-def list_tests(selector: TestSelector) -> None:
-    test_cases = selector.select()
-    print(f"{len(test_cases)} test case(s) selected")
-    for test_case in test_cases:
-        print(f"- {test_case.name}: {test_case.description}")
-
-
-def run_tests(
-    ctx: TestContext, device: NitrokeyTrussedBase, selector: TestSelector
-) -> bool:
-    test_cases = selector.select()
-    if not test_cases:
-        raise CliException("No test cases selected", support_hint=False)
-
-    results = []
-
-    local_print("")
-    local_print(f"Running tests for {device.name} at {device.path}")
-    local_print("")
-
-    n = len(test_cases)
-    idx_len = len(str(n))
-    name_len = max([len(test_case.name) for test_case in test_cases]) + 2
-    description_len = max([len(test_case.description) for test_case in test_cases]) + 2
-    status_len = max([len(status.name) for status in TestStatus]) + 2
-
-    for (i, test_case) in enumerate(test_cases):
-        try:
-            result = test_case.fn(ctx, device)
-        except Exception:
-            result = TestResult(TestStatus.FAILURE, exc_info=sys.exc_info())
-        results.append(result)
-
-        idx = str(i + 1).rjust(idx_len)
-        name = test_case.name.ljust(name_len)
-        description = test_case.description.ljust(description_len)
-        status = result.status.name.ljust(status_len)
-        msg = ""
-        if result.data:
-            msg = str(result.data)
-        elif result.exc_info[1]:
-            logger.error(
-                f"An exception occured during the execution of the test {test_case.name}:",
-                exc_info=result.exc_info,
-            )
-            msg = str(result.exc_info[1])
-
-        local_print(f"[{idx}/{n}]\t{name}\t{description}\t{status}\t{msg}")
-
-    success = len([result for result in results if result.status == TestStatus.SUCCESS])
-    skipped = len([result for result in results if result.status == TestStatus.SKIPPED])
-    failed = len([result for result in results if result.status == TestStatus.FAILURE])
-    local_print("")
-    local_print(f"{n} tests, {success} successful, {skipped} skipped, {failed} failed")
-
-    return all([result.status != TestStatus.FAILURE for result in results])
